Fixing some Confluict
